career_conversation/requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
requests>=2.25.0
pypdf>=3.0.0
gradio>=5.33.0
